refactor: 게시글 상세페이지 버튼들 컴포넌트화 완료
